+++ sdk/storage/storage-file-share/recordings/node/directoryclient/recording_listfilesanddirectories__with_all_attributes.js
@@ -1,59 +1,57 @@
 let nock = require('nock');
 
-module.exports.hash = "0ba5d11d96baf3c058812eb3399db78b";
+module.exports.hash = "6d46174a2d93c2dd4e45919c7b9e5a19";
 
-module.exports.testInfo = {"uniqueName":{"share":"share162544901424303651","dir":"dir162544901554302397","dir0":"dir0162544901674306002","dir1":"dir1162544901762104700","dir2":"dir2162544901879908426","file0":"file0162544901999203768","file1":"file1162544902118808840","file2":"file2162544902239106310"},"newDate":{}}
+module.exports.testInfo = {"uniqueName":{"share":"share167566673899109627","dir":"dir167566673963409154","dir0":"dir0167566673966804575","dir1":"dir1167566673968408551","dir2":"dir2167566673969806061","file0":"file0167566673971408633","file1":"file1167566673972907464","file2":"file2167566673974309764"},"newDate":{}}
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651')
+  .put('/share167566673899109627')
   .query(true)
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:36:55 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F555F9EBEAC"',
+  '"0x8DB080F9F740071"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'd1550721-001a-0007-5c3e-71dd0f000000',
+  '93b391a6-b01a-00b6-68f8-3983b3000000',
   'x-ms-client-request-id',
-  'aad2a6a2-594f-4035-80a7-51655f24e0c8',
+  'ef5519ad-a881-4283-838d-975ecda560e6',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:36:55 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651/dir162544901554302397')
+  .put('/share167566673899109627/dir167566673963409154')
   .query(true)
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:36:57 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F55605808BB"',
+  '"0x8DB080F9F80864F"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '58794bb4-d01a-0012-783e-71af7f000000',
+  '93b391ae-b01a-00b6-69f8-3983b3000000',
   'x-ms-client-request-id',
-  'e6f56cf9-7236-41ab-91a1-18939ffb0c18',
+  '99aef8f3-e353-43b3-a86a-b077ce7bf727',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'x-ms-file-change-time',
-  '2021-07-05T01:36:57.0788027Z',
+  '2023-02-06T06:58:59.6589135Z',
   'x-ms-file-last-write-time',
-  '2021-07-05T01:36:57.0788027Z',
+  '2023-02-06T06:58:59.6589135Z',
   'x-ms-file-creation-time',
-  '2021-07-05T01:36:57.0788027Z',
+  '2023-02-06T06:58:59.6589135Z',
   'x-ms-file-permission-key',
-  '7735077018765312711*253227986890839374',
+  '9212185477508673717*1658283376881248060',
   'x-ms-file-attributes',
   'Directory',
   'x-ms-file-id',
@@ -63,37 +61,35 @@ nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParam
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Mon, 05 Jul 2021 01:36:56 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651/dir162544901554302397/dir0162544901674306002')
+  .put('/share167566673899109627/dir167566673963409154/dir0167566673966804575')
   .query(true)
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:36:57 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F5560DE6350"',
+  '"0x8DB080F9F83BA31"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '58794bc3-d01a-0012-023e-71af7f000000',
+  '93b391af-b01a-00b6-6af8-3983b3000000',
   'x-ms-client-request-id',
-  'b11205b6-2c44-468f-aa1f-2164e6573f17',
+  '364c7d6d-6eac-44f1-9365-0d7fcf35c235',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'x-ms-file-change-time',
-  '2021-07-05T01:36:57.9593040Z',
+  '2023-02-06T06:58:59.6799025Z',
   'x-ms-file-last-write-time',
-  '2021-07-05T01:36:57.9593040Z',
+  '2023-02-06T06:58:59.6799025Z',
   'x-ms-file-creation-time',
-  '2021-07-05T01:36:57.9593040Z',
+  '2023-02-06T06:58:59.6799025Z',
   'x-ms-file-permission-key',
-  '7735077018765312711*253227986890839374',
+  '9212185477508673717*1658283376881248060',
   'x-ms-file-attributes',
   'Directory',
   'x-ms-file-id',
@@ -103,212 +99,200 @@ nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParam
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Mon, 05 Jul 2021 01:36:57 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651/dir162544901554302397/dir1162544901762104700')
+  .put('/share167566673899109627/dir167566673963409154/dir1167566673968408551')
   .query(true)
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:36:59 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F5561922FEB"',
+  '"0x8DB080F9F85DCC2"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '7965f41d-a01a-0005-6b3e-71e2e0000000',
+  '93b391b0-b01a-00b6-6bf8-3983b3000000',
   'x-ms-client-request-id',
-  '8c206338-cd21-4cc7-ba43-afff22de36e0',
+  '801bbded-aca1-4a68-9544-dbceefa19474',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'x-ms-file-change-time',
-  '2021-07-05T01:36:59.1376363Z',
+  '2023-02-06T06:58:59.6938946Z',
   'x-ms-file-last-write-time',
-  '2021-07-05T01:36:59.1376363Z',
+  '2023-02-06T06:58:59.6938946Z',
   'x-ms-file-creation-time',
-  '2021-07-05T01:36:59.1376363Z',
+  '2023-02-06T06:58:59.6938946Z',
   'x-ms-file-permission-key',
-  '7735077018765312711*253227986890839374',
+  '9212185477508673717*1658283376881248060',
   'x-ms-file-attributes',
   'Directory',
   'x-ms-file-id',
-  '13835093239654252544',
+  '16140971433240035328',
   'x-ms-file-parent-id',
   '13835128424026341376',
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Mon, 05 Jul 2021 01:36:59 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651/dir162544901554302397/dir2162544901879908426')
+  .put('/share167566673899109627/dir167566673963409154/dir2167566673969806061')
   .query(true)
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:37:00 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F556247D104"',
+  '"0x8DB080F9F884D6C"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '29283fba-501a-0022-0b3e-71632e000000',
+  '93b391b1-b01a-00b6-6cf8-3983b3000000',
   'x-ms-client-request-id',
-  'ee6b6244-291a-4b04-92aa-5e32fdc20763',
+  'd22f8f39-e7c6-42ea-b88a-b10fb43500fe',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'x-ms-file-change-time',
-  '2021-07-05T01:37:00.3279620Z',
+  '2023-02-06T06:58:59.7098860Z',
   'x-ms-file-last-write-time',
-  '2021-07-05T01:37:00.3279620Z',
+  '2023-02-06T06:58:59.7098860Z',
   'x-ms-file-creation-time',
-  '2021-07-05T01:37:00.3279620Z',
+  '2023-02-06T06:58:59.7098860Z',
   'x-ms-file-permission-key',
-  '7735077018765312711*253227986890839374',
+  '9212185477508673717*1658283376881248060',
   'x-ms-file-attributes',
   'Directory',
   'x-ms-file-id',
-  '13835163608398430208',
+  '10376363910205800448',
   'x-ms-file-parent-id',
   '13835128424026341376',
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Mon, 05 Jul 2021 01:36:59 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651/dir162544901554302397/file0162544901999203768')
+  .put('/share167566673899109627/dir167566673963409154/file0167566673971408633')
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:37:01 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F5562FEAA70"',
+  '"0x8DB080F9F8ABE14"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'c99a6225-101a-001b-473e-7145f9000000',
+  '93b391b2-b01a-00b6-6df8-3983b3000000',
   'x-ms-client-request-id',
-  '805d54f5-7d5a-4b88-ac9b-b44d533341d8',
+  'c4a3e946-afc9-43bb-b42a-b2ce1b115f09',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'x-ms-file-change-time',
-  '2021-07-05T01:37:01.5262832Z',
+  '2023-02-06T06:58:59.7258772Z',
   'x-ms-file-last-write-time',
-  '2021-07-05T01:37:01.5262832Z',
+  '2023-02-06T06:58:59.7258772Z',
   'x-ms-file-creation-time',
-  '2021-07-05T01:37:01.5262832Z',
+  '2023-02-06T06:58:59.7258772Z',
   'x-ms-file-permission-key',
-  '12331785703007272384*253227986890839374',
+  '13809038870468939698*1658283376881248060',
   'x-ms-file-attributes',
   'Archive',
   'x-ms-file-id',
-  '13835075647468208128',
+  '14988049928633188352',
   'x-ms-file-parent-id',
   '13835128424026341376',
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Mon, 05 Jul 2021 01:37:00 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651/dir162544901554302397/file1162544902118808840')
+  .put('/share167566673899109627/dir167566673963409154/file1167566673972907464')
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:37:02 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F5563B69524"',
+  '"0x8DB080F9F8CB99B"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '1c116c67-d01a-000d-6f3e-71ef2a000000',
+  '93b391b3-b01a-00b6-6ef8-3983b3000000',
   'x-ms-client-request-id',
-  '73025b60-0134-4d5e-8e50-f9d8961b0ce2',
+  '5b3a4d2e-9313-407c-a045-4dd088baea3f',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'x-ms-file-change-time',
-  '2021-07-05T01:37:02.7316004Z',
+  '2023-02-06T06:58:59.7388699Z',
   'x-ms-file-last-write-time',
-  '2021-07-05T01:37:02.7316004Z',
+  '2023-02-06T06:58:59.7388699Z',
   'x-ms-file-creation-time',
-  '2021-07-05T01:37:02.7316004Z',
+  '2023-02-06T06:58:59.7388699Z',
   'x-ms-file-permission-key',
-  '12331785703007272384*253227986890839374',
+  '13809038870468939698*1658283376881248060',
   'x-ms-file-attributes',
   'Archive',
   'x-ms-file-id',
-  '13835146016212385792',
+  '12682206919419494400',
   'x-ms-file-parent-id',
   '13835128424026341376',
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Mon, 05 Jul 2021 01:37:02 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .put('/share162544901424303651/dir162544901554302397/file2162544902239106310')
+  .put('/share167566673899109627/dir167566673963409154/file2167566673974309764')
   .reply(201, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Last-Modified',
-  'Mon, 05 Jul 2021 01:37:03 GMT',
+  'Mon, 06 Feb 2023 06:58:59 GMT',
   'ETag',
-  '"0x8D93F556469296E"',
+  '"0x8DB080F9F8F2A4C"',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '3c203c64-a01a-001a-483e-71a2b5000000',
+  '93b391b4-b01a-00b6-6ff8-3983b3000000',
   'x-ms-client-request-id',
-  'd186798d-4fe6-415b-81f6-c9fd01dde77b',
+  'baea980f-9665-4eab-9d4d-9b6f0fcff7f2',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'x-ms-file-change-time',
-  '2021-07-05T01:37:03.9019374Z',
+  '2023-02-06T06:58:59.7548620Z',
   'x-ms-file-last-write-time',
-  '2021-07-05T01:37:03.9019374Z',
+  '2023-02-06T06:58:59.7548620Z',
   'x-ms-file-creation-time',
-  '2021-07-05T01:37:03.9019374Z',
+  '2023-02-06T06:58:59.7548620Z',
   'x-ms-file-permission-key',
-  '12331785703007272384*253227986890839374',
+  '13809038870468939698*1658283376881248060',
   'x-ms-file-attributes',
   'Archive',
   'x-ms-file-id',
-  '13835110831840296960',
+  '17293892937846882304',
   'x-ms-file-parent-id',
   '13835128424026341376',
   'x-ms-request-server-encrypted',
   'true',
   'Date',
-  'Mon, 05 Jul 2021 01:37:03 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .get('/share162544901424303651/dir162544901554302397')
+  .get('/share167566673899109627/dir167566673963409154')
   .query(true)
-  .reply(200, "﻿<?xml version=\"1.0\" encoding=\"utf-8\"?><EnumerationResults ServiceEndpoint=\"https://fakestorageaccount.file.core.windows.net/\" ShareName=\"share162544901424303651\" DirectoryPath=\"dir162544901554302397\"><DirectoryId>13835128424026341376</DirectoryId><Entries><Directory><Name>dir0162544901674306002</Name><FileId>11529285414812647424</FileId><Properties><CreationTime>2021-07-05T01:36:57.9593040Z</CreationTime><LastAccessTime>2021-07-05T01:36:57.9593040Z</LastAccessTime><LastWriteTime>2021-07-05T01:36:57.9593040Z</LastWriteTime><ChangeTime>2021-07-05T01:36:57.9593040Z</ChangeTime><Last-Modified>Mon, 05 Jul 2021 01:36:57 GMT</Last-Modified><Etag>\"0x8D93F5560DE6350\"</Etag></Properties><Attributes>Directory</Attributes><PermissionKey>7735077018765312711*253227986890839374</PermissionKey></Directory><Directory><Name>dir1162544901762104700</Name><FileId>13835093239654252544</FileId><Properties><CreationTime>2021-07-05T01:36:59.1376363Z</CreationTime><LastAccessTime>2021-07-05T01:36:59.1376363Z</LastAccessTime><LastWriteTime>2021-07-05T01:36:59.1376363Z</LastWriteTime><ChangeTime>2021-07-05T01:36:59.1376363Z</ChangeTime><Last-Modified>Mon, 05 Jul 2021 01:36:59 GMT</Last-Modified><Etag>\"0x8D93F5561922FEB\"</Etag></Properties><Attributes>Directory</Attributes><PermissionKey>7735077018765312711*253227986890839374</PermissionKey></Directory><Directory><Name>dir2162544901879908426</Name><FileId>13835163608398430208</FileId><Properties><CreationTime>2021-07-05T01:37:00.3279620Z</CreationTime><LastAccessTime>2021-07-05T01:37:00.3279620Z</LastAccessTime><LastWriteTime>2021-07-05T01:37:00.3279620Z</LastWriteTime><ChangeTime>2021-07-05T01:37:00.3279620Z</ChangeTime><Last-Modified>Mon, 05 Jul 2021 01:37:00 GMT</Last-Modified><Etag>\"0x8D93F556247D104\"</Etag></Properties><Attributes>Directory</Attributes><PermissionKey>7735077018765312711*253227986890839374</PermissionKey></Directory><File><Name>file0162544901999203768</Name><FileId>13835075647468208128</FileId><Properties><Content-Length>1024</Content-Length><CreationTime>2021-07-05T01:37:01.5262832Z</CreationTime><LastAccessTime>2021-07-05T01:37:01.5262832Z</LastAccessTime><LastWriteTime>2021-07-05T01:37:01.5262832Z</LastWriteTime><ChangeTime>2021-07-05T01:37:01.5262832Z</ChangeTime><Last-Modified>Mon, 05 Jul 2021 01:37:01 GMT</Last-Modified><Etag>\"0x8D93F5562FEAA70\"</Etag></Properties><Attributes>Archive</Attributes><PermissionKey>12331785703007272384*253227986890839374</PermissionKey></File><File><Name>file1162544902118808840</Name><FileId>13835146016212385792</FileId><Properties><Content-Length>1024</Content-Length><CreationTime>2021-07-05T01:37:02.7316004Z</CreationTime><LastAccessTime>2021-07-05T01:37:02.7316004Z</LastAccessTime><LastWriteTime>2021-07-05T01:37:02.7316004Z</LastWriteTime><ChangeTime>2021-07-05T01:37:02.7316004Z</ChangeTime><Last-Modified>Mon, 05 Jul 2021 01:37:02 GMT</Last-Modified><Etag>\"0x8D93F5563B69524\"</Etag></Properties><Attributes>Archive</Attributes><PermissionKey>12331785703007272384*253227986890839374</PermissionKey></File><File><Name>file2162544902239106310</Name><FileId>13835110831840296960</FileId><Properties><Content-Length>1024</Content-Length><CreationTime>2021-07-05T01:37:03.9019374Z</CreationTime><LastAccessTime>2021-07-05T01:37:03.9019374Z</LastAccessTime><LastWriteTime>2021-07-05T01:37:03.9019374Z</LastWriteTime><ChangeTime>2021-07-05T01:37:03.9019374Z</ChangeTime><Last-Modified>Mon, 05 Jul 2021 01:37:03 GMT</Last-Modified><Etag>\"0x8D93F556469296E\"</Etag></Properties><Attributes>Archive</Attributes><PermissionKey>12331785703007272384*253227986890839374</PermissionKey></File></Entries><NextMarker /></EnumerationResults>", [
+  .reply(200, "﻿<?xml version=\"1.0\" encoding=\"utf-8\"?><EnumerationResults ServiceEndpoint=\"https://fakestorageaccount.file.core.windows.net/\" ShareName=\"share167566673899109627\" DirectoryPath=\"dir167566673963409154\"><DirectoryId>13835128424026341376</DirectoryId><Entries><Directory><Name>dir0167566673966804575</Name><FileId>11529285414812647424</FileId><Properties><CreationTime>2023-02-06T06:58:59.6799025Z</CreationTime><LastAccessTime>2023-02-06T06:58:59.6799025Z</LastAccessTime><LastWriteTime>2023-02-06T06:58:59.6799025Z</LastWriteTime><ChangeTime>2023-02-06T06:58:59.6799025Z</ChangeTime><Last-Modified>Mon, 06 Feb 2023 06:58:59 GMT</Last-Modified><Etag>\"0x8DB080F9F83BA31\"</Etag></Properties><Attributes>Directory</Attributes><PermissionKey>9212185477508673717*1658283376881248060</PermissionKey></Directory><Directory><Name>dir1167566673968408551</Name><FileId>16140971433240035328</FileId><Properties><CreationTime>2023-02-06T06:58:59.6938946Z</CreationTime><LastAccessTime>2023-02-06T06:58:59.6938946Z</LastAccessTime><LastWriteTime>2023-02-06T06:58:59.6938946Z</LastWriteTime><ChangeTime>2023-02-06T06:58:59.6938946Z</ChangeTime><Last-Modified>Mon, 06 Feb 2023 06:58:59 GMT</Last-Modified><Etag>\"0x8DB080F9F85DCC2\"</Etag></Properties><Attributes>Directory</Attributes><PermissionKey>9212185477508673717*1658283376881248060</PermissionKey></Directory><Directory><Name>dir2167566673969806061</Name><FileId>10376363910205800448</FileId><Properties><CreationTime>2023-02-06T06:58:59.7098860Z</CreationTime><LastAccessTime>2023-02-06T06:58:59.7098860Z</LastAccessTime><LastWriteTime>2023-02-06T06:58:59.7098860Z</LastWriteTime><ChangeTime>2023-02-06T06:58:59.7098860Z</ChangeTime><Last-Modified>Mon, 06 Feb 2023 06:58:59 GMT</Last-Modified><Etag>\"0x8DB080F9F884D6C\"</Etag></Properties><Attributes>Directory</Attributes><PermissionKey>9212185477508673717*1658283376881248060</PermissionKey></Directory><File><Name>file0167566673971408633</Name><FileId>14988049928633188352</FileId><Properties><Content-Length>1024</Content-Length><CreationTime>2023-02-06T06:58:59.7258772Z</CreationTime><LastAccessTime>2023-02-06T06:58:59.7258772Z</LastAccessTime><LastWriteTime>2023-02-06T06:58:59.7258772Z</LastWriteTime><ChangeTime>2023-02-06T06:58:59.7258772Z</ChangeTime><Last-Modified>Mon, 06 Feb 2023 06:58:59 GMT</Last-Modified><Etag>\"0x8DB080F9F8ABE14\"</Etag></Properties><Attributes>Archive</Attributes><PermissionKey>13809038870468939698*1658283376881248060</PermissionKey></File><File><Name>file1167566673972907464</Name><FileId>12682206919419494400</FileId><Properties><Content-Length>1024</Content-Length><CreationTime>2023-02-06T06:58:59.7388699Z</CreationTime><LastAccessTime>2023-02-06T06:58:59.7388699Z</LastAccessTime><LastWriteTime>2023-02-06T06:58:59.7388699Z</LastWriteTime><ChangeTime>2023-02-06T06:58:59.7388699Z</ChangeTime><Last-Modified>Mon, 06 Feb 2023 06:58:59 GMT</Last-Modified><Etag>\"0x8DB080F9F8CB99B\"</Etag></Properties><Attributes>Archive</Attributes><PermissionKey>13809038870468939698*1658283376881248060</PermissionKey></File><File><Name>file2167566673974309764</Name><FileId>17293892937846882304</FileId><Properties><Content-Length>1024</Content-Length><CreationTime>2023-02-06T06:58:59.7548620Z</CreationTime><LastAccessTime>2023-02-06T06:58:59.7548620Z</LastAccessTime><LastWriteTime>2023-02-06T06:58:59.7548620Z</LastWriteTime><ChangeTime>2023-02-06T06:58:59.7548620Z</ChangeTime><Last-Modified>Mon, 06 Feb 2023 06:58:59 GMT</Last-Modified><Etag>\"0x8DB080F9F8F2A4C\"</Etag></Properties><Attributes>Archive</Attributes><PermissionKey>13809038870468939698*1658283376881248060</PermissionKey></File></Entries><NextMarker /></EnumerationResults>", [
   'Transfer-Encoding',
   'chunked',
   'Content-Type',
@@ -316,154 +300,138 @@ nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParam
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '7965f42b-a01a-0005-6f3e-71e2e0000000',
+  '93b391b5-b01a-00b6-70f8-3983b3000000',
   'x-ms-client-request-id',
-  'd9f14a82-ab68-466c-afaf-b073300401a0',
+  '383c7356-a6aa-42c6-8c54-7a1a65f6e34e',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Access-Control-Expose-Headers',
   'x-ms-request-id,x-ms-client-request-id,Server,x-ms-version,Content-Type,Content-Length,Date,Transfer-Encoding',
   'Access-Control-Allow-Origin',
   '*',
   'Date',
-  'Mon, 05 Jul 2021 01:37:04 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/share162544901424303651/dir162544901554302397/file0162544901999203768')
+  .delete('/share167566673899109627/dir167566673963409154/file0167566673971408633')
   .reply(202, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '8f44c0fa-201a-0014-043e-71ee4e000000',
+  '93b391b6-b01a-00b6-71f8-3983b3000000',
   'x-ms-client-request-id',
-  '124c0aab-6fd5-42a7-8fa8-eb1a2096cc1a',
+  '3fbf6ed4-42d6-4684-b33c-fc1cffac9394',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:37:06 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/share162544901424303651/dir162544901554302397/file1162544902118808840')
+  .delete('/share167566673899109627/dir167566673963409154/file1167566673972907464')
   .reply(202, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'd2b1e341-701a-0031-243e-71506f000000',
+  '93b391b7-b01a-00b6-72f8-3983b3000000',
   'x-ms-client-request-id',
-  '5218d437-96b7-4b99-8700-63eed914551e',
+  '3af4a130-ae27-43a7-a5db-780f7d2f4d38',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:37:07 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/share162544901424303651/dir162544901554302397/file2162544902239106310')
+  .delete('/share167566673899109627/dir167566673963409154/file2167566673974309764')
   .reply(202, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'bc2c3dac-801a-0016-5e3e-71d1a1000000',
+  '93b391b8-b01a-00b6-73f8-3983b3000000',
   'x-ms-client-request-id',
-  'b32d3407-b217-4c70-83de-e7d075b9954a',
+  'ddf79141-8bdf-4a38-b4cc-75fe43f704d3',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:37:08 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/share162544901424303651/dir162544901554302397/dir0162544901674306002')
+  .delete('/share167566673899109627/dir167566673963409154/dir0167566673966804575')
   .query(true)
   .reply(202, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'f0d220d0-b01a-0006-7e3e-713a43000000',
+  '93b391b9-b01a-00b6-74f8-3983b3000000',
   'x-ms-client-request-id',
-  '26763e67-26f7-41e8-b7b3-1a42f16d70f5',
+  '7d58eff3-6b62-4b55-b4fc-3b14a91a18e4',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:37:09 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/share162544901424303651/dir162544901554302397/dir1162544901762104700')
+  .delete('/share167566673899109627/dir167566673963409154/dir1167566673968408551')
   .query(true)
   .reply(202, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '927e9cd7-e01a-001d-4c3e-7104c8000000',
+  '93b391ba-b01a-00b6-75f8-3983b3000000',
   'x-ms-client-request-id',
-  'bfc25ffe-dc0f-4144-8ffe-321d4cb8de6c',
+  'de40aaf2-750a-423f-9b17-14f6f23e0990',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:37:11 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/share162544901424303651/dir162544901554302397/dir2162544901879908426')
+  .delete('/share167566673899109627/dir167566673963409154/dir2167566673969806061')
   .query(true)
   .reply(202, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  'f998818f-301a-0029-503e-71b647000000',
+  '93b391bb-b01a-00b6-76f8-3983b3000000',
   'x-ms-client-request-id',
-  '042186c9-4999-4064-9f62-a0e15df5b4b2',
+  '5a498b2b-a346-4812-b707-4d0c5d04065d',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:37:12 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);
 
 nock('https://fakestorageaccount.file.core.windows.net:443', {"encodedQueryParams":true})
-  .delete('/share162544901424303651')
+  .delete('/share167566673899109627')
   .query(true)
   .reply(202, "", [
-  'Transfer-Encoding',
-  'chunked',
+  'Content-Length',
+  '0',
   'Server',
   'Windows-Azure-File/1.0 Microsoft-HTTPAPI/2.0',
   'x-ms-request-id',
-  '2ccaed12-c01a-000e-563e-713789000000',
+  '93b391bc-b01a-00b6-77f8-3983b3000000',
   'x-ms-client-request-id',
-  '3cf0fe7b-9f2c-4389-b351-100f8d3db0d4',
+  '8dfd827d-3c4e-4cdc-a572-1d387f9992e5',
   'x-ms-version',
-  '2020-10-02',
+  '2021-12-02',
   'Date',
-  'Mon, 05 Jul 2021 01:37:13 GMT',
-  'Connection',
-  'close'
+  'Mon, 06 Feb 2023 06:58:59 GMT'
 ]);