+++ sdk/storage/storage-file-share/src/generatedModels.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 // explicitly exporting types that we need.
+import * as coreHttp from "@azure/core-http";
 
 export {
   CopyStatusType,
@@ -9,9 +10,6 @@ export {
   DirectoryCreateResponse,
   DirectoryDeleteResponse,
   DirectoryGetPropertiesResponse,
-  DirectoryItem,
-  DirectoryListFilesAndDirectoriesSegmentResponse,
-  DirectoryListHandlesResponse,
   DirectorySetMetadataResponse,
   DirectorySetPropertiesResponse,
   FileAbortCopyResponse,
@@ -22,9 +20,7 @@ export {
   FileDownloadResponse as FileDownloadResponseModel,
   FileGetPropertiesResponse,
   FileGetRangeListHeaders,
-  FileItem,
   FileLastWrittenMode,
-  FileListHandlesResponse,
   FileServiceProperties,
   FileSetMetadataResponse,
   FileStartCopyResponse,
@@ -32,7 +28,6 @@ export {
   FileUploadRangeFromURLResponse,
   FileUploadRangeResponse,
   PermissionCopyModeType,
-  HandleItem,
   ListSharesIncludeType,
   FileRange as RangeModel,
   ServiceGetPropertiesResponse,
@@ -56,10 +51,7 @@ export {
   DirectoryCreateHeaders,
   DirectoryDeleteHeaders,
   DirectoryGetPropertiesHeaders,
-  FilesAndDirectoriesListSegment,
-  ListFilesAndDirectoriesSegmentResponse,
   DirectoryListFilesAndDirectoriesSegmentHeaders,
-  ListHandlesResponse,
   DirectoryRenameHeaders,
   DirectoryRenameResponse,
   DirectoryListHandlesHeaders,
@@ -115,7 +107,14 @@ export {
   ShareRootSquash,
 } from "./generated/src/models";
 
-import { ShareSetPropertiesResponse, ShareSetPropertiesHeaders } from "./generated/src/models";
+import {
+  ShareSetPropertiesResponse,
+  ShareSetPropertiesHeaders,
+  FileProperty,
+  DirectoryListFilesAndDirectoriesSegmentHeaders,
+  FileListHandlesHeaders,
+  DirectoryListHandlesHeaders,
+} from "./generated/src/models";
 
 /**
  * Contains response data for the setQuota operation.
@@ -126,3 +125,116 @@ export type ShareSetQuotaResponse = ShareSetPropertiesResponse;
  * Defines headers for setQuota operation.
  */
 export type ShareSetQuotaHeaders = ShareSetPropertiesHeaders;
+
+/** A listed file item. */
+export interface FileItem {
+  name: string;
+  fileId?: string;
+  /** File properties. */
+  properties: FileProperty;
+  attributes?: string;
+  permissionKey?: string;
+}
+
+/** A listed directory item. */
+export interface DirectoryItem {
+  name: string;
+  fileId?: string;
+  /** File properties. */
+  properties?: FileProperty;
+  attributes?: string;
+  permissionKey?: string;
+}
+
+/** Abstract for entries that can be listed from Directory. */
+export interface FilesAndDirectoriesListSegment {
+  directoryItems: DirectoryItem[];
+  fileItems: FileItem[];
+}
+
+/** An enumeration of directories and files. */
+export interface ListFilesAndDirectoriesSegmentResponse {
+  serviceEndpoint: string;
+  shareName: string;
+  shareSnapshot?: string;
+  directoryPath: string;
+  prefix: string;
+  marker?: string;
+  maxResults?: number;
+  /** Abstract for entries that can be listed from Directory. */
+  segment: FilesAndDirectoriesListSegment;
+  continuationToken: string;
+  directoryId?: string;
+}
+
+/** Contains response data for the listFilesAndDirectoriesSegment operation. */
+export type DirectoryListFilesAndDirectoriesSegmentResponse =
+  DirectoryListFilesAndDirectoriesSegmentHeaders &
+    ListFilesAndDirectoriesSegmentResponse & {
+      /** The underlying HTTP response. */
+      _response: coreHttp.HttpResponse & {
+        /** The response body as text (string format) */
+        bodyAsText: string;
+
+        /** The response body as parsed JSON or XML */
+        parsedBody: ListFilesAndDirectoriesSegmentResponse;
+        /** The parsed HTTP response headers. */
+        parsedHeaders: DirectoryListFilesAndDirectoriesSegmentHeaders;
+      };
+    };
+
+/** A listed Azure Storage handle item. */
+export interface HandleItem {
+  /** XSMB service handle ID */
+  handleId: string;
+  /** File or directory name including full path starting from share root */
+  path: string;
+  /** FileId uniquely identifies the file or directory. */
+  fileId: string;
+  /** ParentId uniquely identifies the parent directory of the object. */
+  parentId?: string;
+  /** SMB session ID in context of which the file handle was opened */
+  sessionId: string;
+  /** Client IP that opened the handle */
+  clientIp: string;
+  /** Time when the session that previously opened the handle has last been reconnected. (UTC) */
+  openTime: Date;
+  /** Time handle was last connected to (UTC) */
+  lastReconnectTime?: Date;
+}
+
+/** An enumeration of handles. */
+export interface ListHandlesResponse {
+  handleList?: HandleItem[];
+  continuationToken: string;
+}
+
+/** Contains response data for the listHandles operation. */
+export type FileListHandlesResponse = FileListHandlesHeaders &
+  ListHandlesResponse & {
+    /** The underlying HTTP response. */
+    _response: coreHttp.HttpResponse & {
+      /** The response body as text (string format) */
+      bodyAsText: string;
+
+      /** The response body as parsed JSON or XML */
+      parsedBody: ListHandlesResponse;
+      /** The parsed HTTP response headers. */
+      parsedHeaders: FileListHandlesHeaders;
+    };
+  };
+
+/** Contains response data for the listHandles operation. */
+export type DirectoryListHandlesResponse = DirectoryListHandlesHeaders &
+  ListHandlesResponse & {
+    /** The underlying HTTP response. */
+    _response: coreHttp.HttpResponse & {
+      /** The response body as text (string format) */
+      bodyAsText: string;
+
+      /** The response body as parsed JSON or XML */
+      parsedBody: ListHandlesResponse;
+      /** The parsed HTTP response headers. */
+      parsedHeaders: DirectoryListHandlesHeaders;
+    };
+  };